pyproject.toml
[project]
name = "citation-workflow"
version = "0.1.0"
description = "CLI for extracting text from PDFs, finding citations, and scoring events with LLM"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bookwyrm>=0.1.7",
    "click>=8.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
citation-workflow = "citation_workflow.main:main"
